backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.29
requests>=2.31
python-dotenv>=1.0
PyPDF2>=3.0
bcrypt>=4.1
PyJWT>=2.8
python-multipart>=0.0.9
email-validator>=2.2
pydantic>=2.6
pydantic[email]>=2.6
pinecone>=6.0.0
python-docx>=0.8.11






